Added test plant with stages, fixed issues with block replacement/grow block being updated, added workaround for chunk not loading when player spawns into unloaded chunk

+++ src/me/kosinkadink/performantplants/util/TimeHelper.java
@@ -11,7 +11,7 @@ public class TimeHelper {
     }
 
     public static long millisToTicks(long millis) {
-        return (millis/1000) * 20;
+        return (long)((millis/1000.0) * 20);
     }
 
 }
